chore(deps): migrate rusqlite 0.31 -> 0.39 (#35)
Replaces the closed #22.

## Summary
- Bumps \`rusqlite\` from \`0.31\` to \`0.39\` (workspace dep,
\`bundled\` feature kept).

## API delta (only one actually affected us)
- **v0.38 removed default \`ToSql\`/\`FromSql\` impls for \`u64\` /
\`usize\`** (rusqlite #1732). \`alert_counts_by_severity\` used
\`row.get::<_, u64>(1)?\` on \`COUNT(*)\`, which no longer compiles.
Fixed by reading as \`i64\` and casting to \`u64\` — SQLite stores
integers as i64 natively, \`COUNT(*)\` is non-negative, cast is lossless
for realistic row counts.

## Breaking changes audited but not applicable
- **v0.35** \`Connection::execute\` / \`prepare\` reject trailing /
multi-statement SQL → we already use \`execute_batch\` for schema
bootstrap, all other sites are single-statement.
- **v0.34** \`BindIndex\` / \`ValueRef\` error changes → we only use
positional params and typed \`row.get\`.
- **v0.32** \`SQLITE_OPEN_EXRESCODE\` becomes default → transparent.
- **v0.38** min SQLite 3.34.1 → not a concern under \`bundled\`.

**\`libsqlite3-sys\` caveats:** none. The \`bundled\` feature vendors
SQLite, no system-library dep changes on any platform.

## Files touched
- \`Cargo.toml\`
- \`rust/arqenor-store/src/sqlite_store.rs\`

## Verified
- \`cargo check --workspace --exclude arqenor-ebpf --all-targets\` ✅
- \`cargo clippy --workspace --exclude arqenor-ebpf --all-targets -- -D
warnings\` ✅
- \`cargo fmt --all --check\` ✅

+++ rust/arqenor-store/src/sqlite_store.rs
@@ -146,9 +146,12 @@ impl SqliteStore {
         let mut stmt = self.conn.prepare(
             "SELECT severity, COUNT(*) FROM alerts GROUP BY severity ORDER BY COUNT(*) DESC",
         )?;
+        // rusqlite 0.38+ removed the default `FromSql` impl for `u64`/`usize`
+        // (see rusqlite CHANGELOG v0.38). SQLite's COUNT(*) is representable as
+        // `i64`; we cast back to `u64` for the public API (counts are non-negative).
         let rows = stmt
             .query_map([], |row| {
-                Ok((row.get::<_, String>(0)?, row.get::<_, u64>(1)?))
+                Ok((row.get::<_, String>(0)?, row.get::<_, i64>(1)? as u64))
             })?
             .filter_map(|r| r.ok())
             .collect();
